+++ NEWS.md
@@ -1,5 +1,11 @@
 # MsBackendMetabolomicsWorkbench 0.1
 
+## Changes in 0.1.7
+
+- Add `fileName` parameter to `backendInitialize`.
+- Remove `stringr` import.
+- Update tests.
+
 ## Changes in 0.1.6
 
 - Update license to GPL-3.